feat: Update and install Webpack depencies. Migrate configs to Webpack 5 version and type: module package
