+++ .eslintignore
@@ -1,3 +1,4 @@
 node_modules
 dist
+*.config.cjs
 *.config.js
