+++ package.json
@@ -1,30 +1,30 @@
 {
+    "private": true,
     "name": "news-js",
     "type": "module",
     "version": "1.0.0",
-    "description": "",
-    "main": "webpack.base.config.js",
+    "main": "webpack.base.config.cjs",
     "scripts": {
-        "start": "webpack serve --open --config ./webpack.config.js --env mode=dev",
-        "build": "webpack --config ./webpack.config.js --env mode=prod",
+        "start": "webpack serve --open --config ./webpack.config.cjs --env mode=dev",
+        "build": "webpack --config ./webpack.config.cjs --env mode=prod",
         "lint": "eslint . --ext .ts",
         "tsc-dev": "tsc --watch",
         "format": "prettier --write 'src/**/*.{js,ts,json}'"
     },
-    "repository": {
-        "type": "git",
-        "url": ""
-    },
-    "keywords": [],
-    "author": "Aleh Serhiyenia",
-    "license": "ISC",
-    "homepage": "",
     "devDependencies": {
         "@typescript-eslint/eslint-plugin": "^5.59.11",
         "@typescript-eslint/parser": "^5.59.11",
+        "css-loader": "^6.8.1",
         "eslint": "^8.42.0",
         "eslint-config-prettier": "^8.8.0",
+        "html-webpack-plugin": "^5.5.3",
         "prettier": "^2.8.8",
-        "typescript": "^5.1.3"
+        "style-loader": "^3.3.3",
+        "ts-loader": "^9.4.3",
+        "typescript": "^5.1.3",
+        "webpack": "^5.86.0",
+        "webpack-cli": "^5.1.4",
+        "webpack-dev-server": "^4.15.1",
+        "webpack-merge": "^5.9.0"
     }
 }
