+++ src/components/app/app.js
@@ -1,5 +1,5 @@
-import AppController from '../controller/controller';
-import { AppView } from '../view/appView';
+import AppController from '../controller/controller.js';
+import { AppView } from '../view/appView.js';
 
 class App {
   constructor() {
