+++ src/index.js
@@ -1,4 +1,4 @@
-import App from './components/app/app';
+import App from './components/app/app.js';
 import './global.css';
 
 const app = new App();
