+++ webpack.config.cjs
@@ -1,0 +1,39 @@
+/* eslint-env node */
+
+const path = require('path');
+const { merge } = require('webpack-merge');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+const baseConfig = {
+  entry: path.resolve(__dirname, './src/index.js'),
+  mode: 'development',
+  module: {
+    rules: [
+      {
+        test: /\.css$/i,
+        use: ['style-loader', 'css-loader'],
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.js'],
+  },
+  output: {
+    clean: true,
+    filename: 'index.js',
+    path: path.resolve(__dirname, '../dist'),
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: path.resolve(__dirname, './src/index.html'),
+      filename: 'index.html',
+    }),
+  ],
+};
+
+module.exports = ({ mode }) => {
+  const isProductionMode = mode === 'prod';
+  const envConfig = isProductionMode ? require('./webpack.prod.config.cjs') : require('./webpack.dev.config');
+
+  return merge(baseConfig, envConfig);
+};
