+++ webpack.dev.config.cjs
@@ -1,0 +1,11 @@
+const path = require('path');
+
+module.exports = {
+  mode: 'development',
+  devtool: 'inline-source-map',
+  devServer: {
+    static: {
+      directory: path.resolve(__dirname, '../dist'),
+    },
+  },
+};
