--- webpack.dev.config.js
@@ -1,9 +1,0 @@
-const path = require('path');
-
-module.exports = {
-    mode: 'development',
-    devtool: 'inline-source-map',
-    devServer: {
-        contentBase: path.resolve(__dirname, '../dist'),
-    },
-};
